test: move the opt-in invariant into each bucket suite
The R2 cases live in the R2 Bucket suite and the S3 cases in the S3
Bucket suite, next to the live lifecycle tests they harden, instead of a
separate cross-cloud file.
